refactor(fn): moved separate directives into one base class and removed many files in favor of more flat structure. reverted prettier config
