--- libs/fn/src/lib/list/list-item-actions/list-item-actions.directive.spec.ts
@@ -1,8 +1,0 @@
-import { ListItemActionsDirective } from './list-item-actions.directive';
-
-describe('ListItemActionsDirective', () => {
-  it('should create an instance', () => {
-    const directive = new ListItemActionsDirective();
-    expect(directive).toBeTruthy();
-  });
-});
