--- libs/fn/src/lib/list/list-item-actions/list-item-actions.directive.ts
@@ -1,17 +1,0 @@
-import { Directive, HostListener, OnInit, TemplateRef } from '@angular/core';
-
-@Directive({
-    selector: '[fnListItemActions]'
-})
-export class ListItemActionsDirective implements OnInit {
-    constructor(readonly templateRef: TemplateRef<any>) {}
-
-    @HostListener('click', ['$event'])
-    onHostClick($event: MouseEvent): void {
-        console.log({ $event });
-    }
-
-    ngOnInit(): void {
-        console.log('test');
-    }
-}
