--- libs/fn/src/lib/list/list-item-end/list-item-actions.directive.spec.ts
@@ -1,8 +1,0 @@
-import { ListItemEndDirective } from './list-item-end.directive';
-
-describe('ListItemActionsDirective', () => {
-  it('should create an instance', () => {
-    const directive = new ListItemEndDirective();
-    expect(directive).toBeTruthy();
-  });
-});
