--- libs/fn/src/lib/list/list-item-end/list-item-end.directive.ts
@@ -1,8 +1,0 @@
-import { Directive, TemplateRef } from '@angular/core';
-
-@Directive({
-    selector: '[fnListItemEnd]'
-})
-export class ListItemEndDirective {
-    constructor(readonly templateRef: TemplateRef<any>) {}
-}
