--- libs/fn/src/lib/list/list-item-title/list-item-title.directive.spec.ts
@@ -1,8 +1,0 @@
-import { ListItemTitleDirective } from './list-item-title.directive';
-
-describe('ListItemTitleDirective', () => {
-    it('should create an instance', () => {
-        const directive = new ListItemTitleDirective();
-        expect(directive).toBeTruthy();
-    });
-});
