+++ libs/fn/src/lib/list/list-item/list-item.component.html
@@ -4,7 +4,7 @@
 
 <ng-content></ng-content>
 
-<div class="fn-list__end-container" *ngIf="actionsDirective || endDirective">
+<div class="fn-list__end-container" *ngIf="actionsDirective || endDirective" (click)="$event.stopPropagation()">
     <ng-container *ngTemplateOutlet="endDirective?.templateRef"></ng-container>
     <ng-container *ngTemplateOutlet="actionsDirective?.templateRef"></ng-container>
 </div>
