+++ libs/fn/src/lib/list/list.module.ts
@@ -5,9 +5,7 @@ import { ListItemComponent } from './list-item/list-item.component';
 import { ListGroupHeaderComponent } from './list-group-header/list-group-header.component';
 import { SelectableItemDirective } from './selectable-item/selectable-item.directive';
 import { SelectableListDirective } from './selectable-list/selectable-list.directive';
-import { ListItemTitleDirective } from './list-item-title/list-item-title.directive';
-import { ListItemActionsDirective } from './list-item-actions/list-item-actions.directive';
-import { ListItemEndDirective } from './list-item-end/list-item-end.directive';
+import { ListItemTitleDirective, ListItemActionsDirective, ListItemEndDirective } from './structural-directives';
 
 @NgModule({
     imports: [CommonModule],
