+++ libs/fn/src/lib/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './utils.module';
 export * from './decorators/coerceBoolean';
+export * from './template-ref.directive';
