+++ libs/fn/src/lib/utils/template-ref.directive.ts
@@ -1,0 +1,6 @@
+import { Directive, TemplateRef } from '@angular/core';
+
+@Directive()
+export class TemplateRefDirective<ContextType = any> {
+    constructor(readonly templateRef: TemplateRef<ContextType>) {}
+}
